adding helper methods to util.ts file
